First round of refactoring to tests
